pyproject.toml
[project]
name = "template_project"
requires-python = ">=3.13"
description = "template project"
version = "1.0.0"
dependencies = [
    "uuid-utils==0.11.1",
    "adaptix==3.0.0b11",
    "fastapi==0.119.0",
    "uvicorn==0.37.0",
    "dishka==1.7.2",
    "argon2-cffi==23.1.0",
    "cryptography==46.0.3",
]
